+++ src/Serilog.Sinks.Http/Sinks/Http/Private/Network/DefaultHttpClient.cs
@@ -11,13 +11,13 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-
+#pragma warning disable CS1591 // Missing XML comment for publicly visible type or member
 using System.Net.Http;
 using System.Threading.Tasks;
 
 namespace Serilog.Sinks.Http.Private.Network
 {
-    internal class DefaultHttpClient : IHttpClient
+    public class DefaultHttpClient : IHttpClient
     {
         private readonly HttpClient client;
 
@@ -31,3 +31,4 @@ namespace Serilog.Sinks.Http.Private.Network
             client.Dispose();
     }
 }
+#pragma warning restore CS1591 // Missing XML comment for publicly visible type or member