+++ src/Serilog.Sinks.Http/Sinks/Http/Private/Time/PortableTimer.cs
@@ -11,14 +11,14 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-
+#pragma warning disable CS1591 // Missing XML comment for publicly visible type or member
 using System;
 using System.Threading;
 using System.Threading.Tasks;
 
 namespace Serilog.Sinks.Http.Private.Time
 {
-    internal class PortableTimer : IDisposable
+    public class PortableTimer : IDisposable
     {
         private readonly object stateLock = new object();
         private readonly Func<Task> onTick;
@@ -109,3 +109,4 @@ namespace Serilog.Sinks.Http.Private.Time
         }
     }
 }
+#pragma warning restore CS1591 // Missing XML comment for publicly visible type or member